Allow sharing modules dir between many bots, compile cleanly with -Wextra
The module data directory is now /data instead of /modules/data. This
allows /modules to be a symlink to a shared location while each bot can
have its own set of /data files.
This directory will be auto migrated to the new location if possible.

Added modules.include and modules.exclude files. These allow controling
which modules in the /modules dir will be loaded, instead of the
previous behaviour of always loading them all. Again, this is useful if
/modules is being shared.

Compile with -Wextra, ignoring a few things. The ignores made CFLAGS very
long, so it has been moved to the src/CFLAGS file, and included using
gcc/clang's @ option.

Also tweak a few things that were brought up by clang --analyze.