mod_psa: add regex trigger support + alias interop
The regex format is POSIX extended (via libc's regcomp).

If you were previously using the '*' trigger feature, to send on any
message then you should probably change that to '.'

To make the psa invoke an alias, just set the message to be the alias
name (including the !) e.g:
   !psa+ test 1m !keyboard
will invoke the !keyboard alias every minute.