split messages differently
on_filter is now only called once per message (before it is split)
previously it would be called multiple times with the same ID for each
split.

This was a problem for mod_extadmin which didn't expect the same ID more
than once, and would fail to filter the 2nd+ splits.